chore(26088): a bit of renaming
