+++ hivemq-edge/src/frontend/src/components/rjsf/MqttTransformation/hooks/useMappingValidation.tsx
@@ -1,7 +1,7 @@
-import { MappingValidation, OutwardSubscription } from '@/modules/Subscriptions/types.ts'
+import { MappingValidation, OutwardMapping } from '@/modules/Mappings/types.ts'
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
-export const useMappingValidation = (_item: OutwardSubscription) => {
+export const useMappingValidation = (_item: OutwardMapping) => {
   const status: MappingValidation = { status: 'error', errors: [] }
   return status
 }
